admin/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0



elasticsearch>=6.0.0,<7.0.0
elasticsearch-dsl>=6.0.0,<7.0.0
django-elasticsearch-dsl>=6.0.0,<7.0.0
django-elasticsearch-dsl-drf==0.18